# docs/swagger.yaml
basePath: /v1
definitions:
...
  dto.DeleteMessageForEveryoneRequest:
    properties:
      fromMe:
        type: boolean
      id:
        type: string
      participant:
        type: string
      remoteJid:
        type: string
    required:
    - id
    - remoteJid
    type: object
...
  dto.EditMessageRequest:
    properties:
      id:
        type: string
      newMessage:
        type: string
      remoteJid:
        type: string
    required:
    - id
    - newMessage
    - remoteJid
    type: object
...
  dto.NumberExistsRequest:
    properties:
      numbers:
        items:
          type: string
        minItems: 1
        type: array
    required:
    - numbers
    type: object
...
  dto.ReadMessagesRequest:
    properties:
      readMessages:
        items:
          $ref: '#/definitions/dto.ReadMessagesRequestItem'
        minItems: 1
        type: array
    required:
    - readMessages
    type: object
  dto.ReadMessagesRequestItem:
    properties:
      id:
        type: string
      remoteJid:
        type: string
      sender:
        description: FromMe    bool   `json:"fromMe"` ignored
        type: string
    required:
    - id
    - remoteJid
    type: object
  dto.ReplyMessageRequest:
    properties:
      messageId:
        type: string
      originalSender:
        type: string
      participant:
        type: string
      remoteJid:
        type: string
      text:
        type: string
    required:
    - messageId
    - remoteJid
    - text
    type: object
...
  dto.SendChatPresenceRequest:
    properties:
      delay:
        maximum: 300000
        minimum: 0
        type: integer
      number:
        type: string
      presence:
        $ref: '#/definitions/dto.SendPresenceRequestPresence'
      type:
        $ref: '#/definitions/dto.SendPresenceRequestType'
    type: object
...
  dto.SendPresenceRequestPresence:
    enum:
    - composing
    - available
    type: string
    x-enum-varnames:
    - PresenceComposing
    - PresenceAvailable
  dto.SendPresenceRequestType:
    enum:
    - text
    - audio
    type: string
    x-enum-varnames:
    - PresenceTypeText
    - PresenceTypeAudio
...
  utils.HTTPErrorResponse:
    properties:
      error:
        type: string
      errorMessage:
        type: string
      message:
        type: string
    type: object
...
  whatsmiau.SendTextResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: Verbeux AI
    url: https://github.com/verbeux-ai/whatsmiau
  description: WhatsMiau - WhatsApp API compatible with Evolution API. Provides instance
    management, messaging, and chat operations over WhatsApp Web.
  license:
    name: MIT
  title: WhatsMiau API
  version: 0.3.2
paths:
...
  /chat/deleteMessageForEveryone/{instance}:
    delete:
      consumes:
      - application/json
      description: Revokes a message in the chat so it is deleted for all participants.
      parameters:
      - description: Instance ID
        in: path
        name: instance
        required: true
        type: string
      - description: Message to revoke
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.DeleteMessageForEveryoneRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Empty object on success
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete message for everyone
      tags:
      - Chat
  /chat/editMessage/{instance}:
    put:
      consumes:
      - application/json
      description: Edits a previously sent message with new text content.
      parameters:
      - description: Instance ID
        in: path
        name: instance
        required: true
        type: string
      - description: Message to edit
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.EditMessageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Empty object on success
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Edit a sent message
      tags:
      - Chat
  /chat/markMessageAsRead/{instance}:
    post:
      consumes:
      - application/json
      description: Marks one or more messages as read in a WhatsApp conversation
      parameters:
      - description: Instance ID
        in: path
        name: instance
        required: true
        type: string
      - description: Messages to mark as read
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ReadMessagesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Empty object on success
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Mark messages as read
      tags:
      - Chat
  /chat/reply/{instance}:
    post:
      consumes:
      - application/json
      description: Sends a text reply to a specific message in a chat.
      parameters:
      - description: Instance ID
        in: path
        name: instance
        required: true
        type: string
      - description: Reply parameters
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ReplyMessageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Reply sent successfully
          schema:
            $ref: '#/definitions/whatsmiau.SendTextResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Reply to a message
      tags:
      - Chat
  /chat/sendPresence/{instance}:
    post:
      consumes:
      - application/json
      description: Sends a presence status (composing/available) to a WhatsApp contact,
        with optional auto-stop delay
      parameters:
      - description: Instance ID
        in: path
        name: instance
        required: true
        type: string
      - description: Presence parameters
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.SendChatPresenceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Empty object on success
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Send chat presence (typing indicator)
      tags:
      - Chat
  /chat/whatsappNumbers/{instance}:
    post:
      consumes:
      - application/json
      description: Checks whether the given phone numbers are registered on WhatsApp
      parameters:
      - description: Instance ID
        in: path
        name: instance
        required: true
        type: string
      - description: Numbers to check
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.NumberExistsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: List of number existence results
          schema:
            items:
              type: object
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Check if numbers exist on WhatsApp
      tags:
      - Chat
...
  /instance/{instance}/chat/deleteMessageForEveryone:
    delete:
      consumes:
      - application/json
      description: Revokes a message in the chat so it is deleted for all participants.
      parameters:
      - description: Instance ID
        in: path
        name: instance
        required: true
        type: string
      - description: Message to revoke
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.DeleteMessageForEveryoneRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Empty object on success
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete message for everyone
      tags:
      - Chat
  /instance/{instance}/chat/editMessage:
    put:
      consumes:
      - application/json
      description: Edits a previously sent message with new text content.
      parameters:
      - description: Instance ID
        in: path
        name: instance
        required: true
        type: string
      - description: Message to edit
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.EditMessageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Empty object on success
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Edit a sent message
      tags:
      - Chat
  /instance/{instance}/chat/presence:
    post:
      consumes:
      - application/json
      description: Sends a presence status (composing/available) to a WhatsApp contact,
        with optional auto-stop delay
      parameters:
      - description: Instance ID
        in: path
        name: instance
        required: true
        type: string
      - description: Presence parameters
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.SendChatPresenceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Empty object on success
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Send chat presence (typing indicator)
      tags:
      - Chat
  /instance/{instance}/chat/read-messages:
    post:
      consumes:
      - application/json
      description: Marks one or more messages as read in a WhatsApp conversation
      parameters:
      - description: Instance ID
        in: path
        name: instance
        required: true
        type: string
      - description: Messages to mark as read
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ReadMessagesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Empty object on success
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Mark messages as read
      tags:
      - Chat
  /instance/{instance}/chat/reply:
    post:
      consumes:
      - application/json
      description: Sends a text reply to a specific message in a chat.
      parameters:
      - description: Instance ID
        in: path
        name: instance
        required: true
        type: string
      - description: Reply parameters
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ReplyMessageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Reply sent successfully
          schema:
            $ref: '#/definitions/whatsmiau.SendTextResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Reply to a message
      tags:
      - Chat
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: apikey
    type: apiKey
swagger: "2.0"